+++ primal-memory/bench/Conversion.hs
@@ -25,13 +25,13 @@ main = do
   mb2 <- allocAlignedMBytes n64
   mb3 <- allocAlignedMBytes n64
   let fp = toForeignPtrMBytes mb3
-  mba <- BA.newAlignedPinnedByteArray (fromCount (n :: Count Word64)) 8
+  mba <- BA.newAlignedPinnedByteArray (unCountBytes (n :: Count Word64)) 8
   ba <- BA.unsafeFreezeByteArray mba
   -- Ensure that arrays are equal by filling them with zeros
   bEq1 <- freezeMBytes =<< callocAlignedMBytes n64
   bEq2 <- freezeMBytes =<< callocAlignedMBytes n64
-  mbaEq1 <- BA.newAlignedPinnedByteArray (fromCount (n :: Count Word64)) 8
-  mbaEq2 <- BA.newAlignedPinnedByteArray (fromCount (n :: Count Word64)) 8
+  mbaEq1 <- BA.newAlignedPinnedByteArray (unCountBytes (n :: Count Word64)) 8
+  mbaEq2 <- BA.newAlignedPinnedByteArray (unCountBytes (n :: Count Word64)) 8
   BA.setByteArray mbaEq1 0 (unCount n64) (0 :: Word64)
   BA.setByteArray mbaEq2 0 (unCount n64) (0 :: Word64)
   baEq1 <- BA.unsafeFreezeByteArray mbaEq1