+++ primal-memory/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for primal-memory
 
-## 0.1.0.0
+## 0.2.0
+
+* Rename `ByteArray` -> `PrimArray`
+* Improve `fromList` conversion
+
+## 0.1.0
 
 * Initial release